+++ config/initializers/mail.rb
@@ -12,6 +12,22 @@ if Rails.env.production?
     #:enable_starttls_auto => true
  }
 
+  # Even in a production environment, we want to check the env var for
+  # our staging setup for testing or other purposes, so real users don't
+  # receive test emails.
+  unless Rails.application.secrets.smtp_override_recipient.nil? || Rails.application.secrets.smtp_override_recipient.empty?
+    class OverrideRecipientInterceptor
+      def delivering_email(message)
+        message.to = [Rails.application.secrets.smtp_override_recipient]
+        message.cc = nil
+        message.bcc = nil
+      end
+    end
+
+    ActionMailer::Base.register_interceptor(OverrideRecipientInterceptor.new)
+  end
+
+
 else
 
   ActionMailer::Base.smtp_settings = {
@@ -30,7 +46,7 @@ else
   # testing or other purposes, so real users don't receive test emails.
   class OverrideRecipientInterceptor
     def delivering_email(message)
-      message.to = Rails.application.secrets.smtp_override_recipient
+      message.to = [Rails.application.secrets.smtp_override_recipient]
       message.cc = nil
       message.bcc = nil
     end
